requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-seed==0.2.2
Faker==6.6.1
idna==2.10
mysqlclient==2.0.3
oauthlib==3.1.0
Pillow==8.1.2
pkg-resources==0.0.0
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
uWSGI==2.0.19.1
